+++ src/index.test.js
@@ -1,6 +1,29 @@
-const { __fetchProjectData, __uuidv4 } = require('./');
+const { __fetchProjectData, __uuidv4, onTrack } = require('./');
 
-// TODO(derek@optimizely.com) Update tests
+let fetchMock;
+
+beforeEach(() => {
+  fetchMock = {
+    fn: () =>
+      new Promise(resolve =>
+        resolve({
+          json: () =>
+            new Promise(resolve =>
+              resolve({
+                accountId: '123456',
+                attributes: [{ id: '9876', key: 'attr_key_one' }],
+                events: [{ id: '5432', key: 'event_key_one' }],
+              })
+            ),
+        })
+      ),
+  };
+});
+
+// TODO(derek@optimizely.com) Add tests
+describe('onTrack', () => {});
+
+// TODO(derek@optimizely.com) Add tests
 describe('__fetchProjectData', () => {
   it('fails as expected with no SDK Key', () => {});
   it('fails as expected with invalid SDK Key', () => {});
